# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estudy",
        "description": "API doc for Estudy",
        "contact": {
            "url": "https://justice-chimobi.vercel.app/"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/"
        }
    ],
    "paths": {
        "/auth/signin": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Logs existing user with credentials",
                "description": "authentication endpoint using JWT token",
                "operationId": "be2d54ca411de1e23a8c053ef38f2e9a",
                "requestBody": {
                    "description": "authenticate if user exists",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "giftowens@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secrets"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Logged in successfully",
                                    "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0OjgwMDAvYXBpL2F1dGgvc2lnbAsIm5iZiI6MTcwNjAwNDI1MCwianRpIjoidE1tUlNHbjg0QlZoQkdYTSIsInN1YiI6IjU1MjJmNjQyLTYyOGQtNDZkZi04NDMyLTNmZTBjMjllNDM5YiIsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.xoKfHEHLZ6vDwIfg-GZQXoMwG9wQGlgs4vDrLTEVwoU",
                                    "user_type": "mentor"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
...
        "/courses/m/create": {
            "post": {
                "tags": [
                    "courses"
                ],
                "summary": "Created course only by mentor",
                "description": "Created course only by mentor",
                "operationId": "9467a528b0e8ee2ccdd3043ca73702d7",
                "requestBody": {
                    "description": "Created courses by mentor only: (Get your videos from return payload at /courses/m/video) and (thumbnail from return payload /courses/m/thumnail)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "estudy course name"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 10
                                    },
                                    "video": {
                                        "type": "string",
                                        "example": [
                                            "https://res.cloudinary.com/estudy/image/upload/v1705789451/yofikr4gyecw04sp5ial.mp4",
                                            "https://res.cloudinary.com/estudy/image/upload/v1705789451/yofikr4gyecw04sp5ial.mp4"
                                        ]
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Web Development"
                                    },
                                    "thumbnail": {
                                        "type": "string",
                                        "example": "https://res.cloudinary.com/estudy/image/upload/v1705789451/yofikr4gyecw04sp5ial.png"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "estudy course description"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": [
                                            "Title 1",
                                            "Title 2"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Course created successfully"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/courses/m/{id}": {
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Get a courses by id example to update",
                "description": "Get a courses by id example to update",
                "operationId": "87d166943319d39322999a930764c094",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "An UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "example": {
                                                "id": "550e8400-e29b-41d4-a716-446655440000",
                                                "name": "estudy",
                                                "price": "20.00",
                                                "category": "mobile development",
                                                "description": "estudy decription",
                                                "thumbnail": "https://res.cloudinary.com/estudy/image/upload/v1705789451/yofikr4gyecw04sp5ial.jpg",
                                                "video": [
                                                    {
                                                        "url": "https://example.com/video1.mp4"
                                                    },
                                                    {
                                                        "url": "https://example.com/video2.mp4"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "courses"
                ],
                "summary": "Mentor delete a created course",
                "description": "Mentor delete a created course",
                "operationId": "cf6a30bec52e3c895b1506dbf176ac61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "An UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "courses"
                ],
                "summary": "Update course created by a mentor",
                "description": "Update course created by a mentor",
                "operationId": "c5f498fbe2559ae87c110cb8c1a94560",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "An UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "description": "Created courses by mentor only: (Get your videos from return payload at /courses/m/video) and (thumbnail from return payload /courses/m/thumnail)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "estudy course name"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 10
                                    },
                                    "video": {
                                        "type": "string",
                                        "example": [
                                            "https://res.cloudinary.com/estudy/image/upload/v1705789451/yofikr4gyecw04sp5ial.mp4",
                                            "https://res.cloudinary.com/estudy/image/upload/v1705789451/yofikr4gyecw04sp5ial.mp4"
                                        ]
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Web Development"
                                    },
                                    "thumbnail": {
                                        "type": "string",
                                        "example": "https://res.cloudinary.com/estudy/image/upload/v1705789451/yofikr4gyecw04sp5ial.png"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "estudy course description"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": [
                                            "Title 1",
                                            "Title 2"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Course updated successfully"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}